requirements.txt
asyncio-nats-client~=0.9.2
fastapi==0.78.0
uvicorn[standard]
gunicorn==20.1.0